+++ de-DE/tutorials/Invoke-GetHelpTutorial.ps1
@@ -6,30 +6,31 @@ Param(
 )
 
 $cmd = 'Get-Help'
-$title = "$($titleStyle)Getting Started with $cmd$reset"
+$title = "$($titleStyle)Erste Schritte mit $cmd$reset"
 
 #region content
 $Intro = @"
 
-{0}$cmd{1} is another of the basic PowerShell cmdlets you need to understand how to use. You
-will be using this command constantly. You should refer to the help often. Each new release
-of PowerShell might introduce new commands, new parameters, new help examples as well as
-correcting errors in documentation.
+{0}$cmd{1} ist eines der grundlegenden PowerShell‑Cmdlets, deren Verwendung Sie verstehen müssen. Sie
+werden diesen Befehl ständig verwenden. Sie sollten häufig in der Hilfe nachsehen. Jede neue
+Version von PowerShell kann neue Befehle, neue Parameter, neue Hilfe‑Beispiele einführen sowie
+Fehler in der Dokumentation korrigieren.
 
-Out of the box, PowerShell ships with a subset of help documentation. Microsoft will also
-periodically update help documentation, although there is no notification mechanism. After
-every PowerShell update and periodically, you should run {0}Update-Help{1}.
+Standardmäßig enthält PowerShell nur einen Teil der Hilfedokumentation. Microsoft aktualisiert
+die Hilfedateien ebenfalls periodisch, obwohl es keinen Benachrichtigungsmechanismus gibt. Nach
+jedem PowerShell‑Update und in regelmäßigen Abständen sollten Sie {0}Update-Help{1} ausführen.
 
 $prompt {0}Update-Help{1}
 
-A few errors are to be expected.
+Einige Fehler sind zu erwarten.
 "@ -f $cmdStyle, $reset
 
 $P1 = @"
 
-While learning PowerShell, don't try to memorize how to use a PowerShell command. Instead,
-learn to {2}discover{1} how to use it. You can use the {0}$cmd{1} cmdlet or helper function, {2}help{1},
-which provides the same information as the cmdlet, but with paged output.
+Beim Erlernen von PowerShell sollten Sie nicht versuchen, sich einzuprägen, wie ein
+PowerShell‑Befehl verwendet wird. Lernen Sie stattdessen, wie Sie seine Verwendung {2}entdecken{1}.
+Sie können das Cmdlet {0}$cmd{1} oder die Hilfsfunktion {2}help{1} verwenden, die dieselben
+Informationen wie das Cmdlet liefert, jedoch seitenweise Ausgabe bietet.
 
 $prompt {3}help{1} {4}Get-Process{1}
 
@@ -37,7 +38,7 @@ $prompt {3}help{1} {4}Get-Process{1}
 
 $P2 = @"
 
-The output headings should be self-explanatory. You can see the command {0}name{1}.
+Die Ausgabeüberschriften sollten selbsterklärend sein. Sie können den Befehlsnamen {0}Name{1} sehen.
 
 NAME
     Get-Process
@@ -45,7 +46,7 @@ NAME
 "@ -f $highLight,$reset
 
 $P3 = @"
-A brief {0}synopsis{1} or explanation about what the command does.
+Eine kurze {0}Kurzbeschreibung{1} oder Erklärung darüber, was der Befehl bewirkt.
 
 SYNOPSIS
     Gets the processes that are running on the local computer.
@@ -53,7 +54,7 @@ SYNOPSIS
 "@ -f $highLight,$reset
 
 $P4 = @"
-The command {0}syntax{1}, which is the same thing you can see with {2}Get-Command{1}.
+Die Befehls{0}syntax{1}, die Sie auch mit {2}Get-Command{1} sehen können.
 
 SYNTAX
     Get-Process [[-Name] <System.String[]>] [-FileVersionInfo] [-Module] [<CommonParameters>]
@@ -72,7 +73,7 @@ SYNTAX
 "@ -f $highLight,$reset,$cmdStyle
 
 $P5 = @"
-And a more detailed {0}description{1}.
+Und eine ausführlichere {0}Beschreibung{1}.
 
 DESCRIPTION
     The `Get-Process` cmdlet gets the processes on a local computer.
@@ -82,8 +83,8 @@ DESCRIPTION
 
 $P6 = @"
 
-Most command help will also include one or more examples. If you want to see only the examples
-you can run:
+Die Hilfe zu den meisten Befehlen enthält in der Regel ein oder mehrere Beispiele. Wenn Sie nur die
+Beispiele sehen möchten, können Sie Folgendes ausführen:
 
 $prompt {3}help{1} {4}Get-Process{1} {5}-examples{1}
 
@@ -108,10 +109,10 @@ SYNOPSIS
 
 $P7 = @"
 
-Finally, you can always get the most up-to-date help by using the {0}-Online{1} parameter. If there
-is an online link, it will open in your default browser. It is not a requirement, but the majority
-of the commands you will be using support this help feature. The online link will be displayed
-under RELATED LINKS:
+Schließlich können Sie jederzeit die aktuellste Hilfe abrufen, indem Sie den Parameter {0}-Online{1}
+verwenden. Wenn ein Online-Link vorhanden ist, wird dieser in Ihrem Standardbrowser geöffnet. Es ist
+nicht erforderlich, ihn zu verwenden, aber die Mehrheit der Befehle, die Sie verwenden werden,
+unterstützt diese Hilfefunktion. Der Online-Link wird unter RELATED LINKS angezeigt:
 
 RELATED LINKS
     {2}Online Version:
@@ -126,7 +127,7 @@ RELATED LINKS
 
 $P8 = @"
 
-It is very important that you understand how to interpret the help syntax.
+Es ist sehr wichtig, dass Sie verstehen, wie man die Hilfesyntax interpretiert.
 
 SYNTAX
     Get-Process [[-Name] <System.String[]>] [-FileVersionInfo] [-Module] [<CommonParameters>]
@@ -134,14 +135,14 @@ SYNTAX
     Get-Process [-FileVersionInfo] -Id <System.Int32[]> [-Module] [<CommonParameters>]
     ...
 
-Each combination of parameters is referred as a {0}parameter set{1}. You cannot mix and match
-parameters between sets. For now, you can ignore CommonParameters.
+Jede Kombination von Parametern wird als {0}Parametersatz{1} bezeichnet. Sie können Parameter nicht
+zwischen verschiedenen Sätzen mischen. Im Moment können Sie CommonParameters ignorieren.
 
 "@ -f $highLight,$reset
 
 $P9 = @"
-Anything you see enclosed in [ ] is considered {0}optional{2}. This means that if you want to use
-the {4}Id{2} parameter, you must specify it because it is not enclosed in [ ].
+Alles, was Sie in [ ] eingeschlossen sehen, gilt als {0}optional{2}. Das bedeutet, wenn Sie den
+{4}Id{2}-Parameter verwenden möchten, müssen Sie ihn angeben, da er nicht in [ ] eingeschlossen ist.
 
 SYNTAX
     Get-Process [[-Name] <System.String[]>] [-FileVersionInfo] [-Module] [<CommonParameters>]
@@ -154,45 +155,49 @@ $prompt {3}Get-Process{2} {4}-Id{2} {5}`$PID{2}
 
 $P10 = @"
 
-{0}`$PID{1} is an automatic variable whose value is the process ID for the current PowerShell session.
+{0}`$PID{1} ist eine automatische Variable, deren Wert die Prozess-ID der aktuellen PowerShell-Sitzung ist.
+
 
 "@ -f $highLight,$reset
 
 $P10a = @"
-Looking at the syntax again, you can use the {1}Name{4} parameter without having to type it because
-the parameter name is enclosed in [ ].
+Wenn Sie die Syntax noch einmal betrachten, können Sie den {1}Name{4}-Parameter verwenden, ohne ihn
+eingeben zu müssen, da der Parametername in [ ] eingeschlossen ist.
 
 SYNTAX
     Get-Process [{2}[{4}-Name{2}]{4} <System.String[]>] [-FileVersionInfo] [-Module]
     [<CommonParameters>]
 
-The syntax tells you that the entire {1}Name{4} parameter is optional, and if you want to use it,
-you don't even need to use the parameter name. This is referred to as a {0}positional{4} parameter.
+Die Syntax zeigt Ihnen, dass der gesamte {1}Name{4}-Parameter optional ist, und wenn Sie ihn verwenden
+möchten, müssen Sie den Parameternamen nicht einmal angeben. Dies wird als {0}positionsabhängiger{4}
+Parameter bezeichnet.
 
-The text within the {2}< >{4} indicates what {0}type{4} of parameter value to use. System.String means
-text. When you see {2}[]{4} as part of the type, this indicates the parameter will accept multiple values
-separated by commas.
+Der Text innerhalb der {2}< >{4} gibt an, welchen {0}Typ{4} von Parameterwert Sie verwenden sollen.
+System.String bedeutet Text. Wenn Sie {2}[]{4} als Teil des Typs sehen, zeigt dies an, dass der Parameter
+mehrere durch Kommas getrennte Werte akzeptiert.
 
 $prompt {3}Get-Process{4} {5}pwsh,sys*{4}
 "@ -f $highLight,$highLight2,$highlight3,$cmdStyle,$reset,$defaultTokenStyle
 
 $P11 = @"
-You can always look at parameter detail in the help. You can specify parameter names using wildcards.
+Sie können jederzeit die Parameterdetails in der Hilfe anzeigen. Parameternamen können mit
+Platzhaltern angegeben werden.
 
 $prompt {0}help{3} {2}Get-Process{3} {1}-Parameter{3} {2}*name{3}
 "@ -f $cmdStyle,$paramStyle,$defaultTokenStyle,$reset
 
 $P12 = @"
-Do you see how these parameters are defined as positional or named? The detail also indicates
-if it is a required parameter. Compare this information to how it is displayed in the help.
+Sehen Sie, wie diese Parameter als positionsabhängig oder benannt definiert sind? Die Detailangaben
+zeigen auch an ob es sich um einen erforderlichen Parameter handelt. Vergleichen Sie diese Informationen
+damit, wie sie in der Hilfe angezeigt werden.
 
     Get-Process [[-Name] <System.String[]>] [-FileVersionInfo] [-Module] [<CommonParameters>]
     Get-Process -Id <System.Int32[]> -IncludeUserName [<CommonParameters>]
     Get-Process [[-Name] <System.String[]>] -IncludeUserName [<CommonParameters>]
     Get-Process -IncludeUserName -InputObject <System.Diagnostics.Process[]> [<CommonParameters>]
 
-If you wanted to see all PowerShell processes and those beginning with {0}sys{1} AND show the
-user name, do you have an idea of what command you would type?
+Wenn Sie alle PowerShell‑Prozesse und diejenigen sehen möchten, die mit {0}sys{1} beginnen UND
+den Benutzernamen anzeigen möchten, haben Sie eine Idee, welchen Befehl Sie eingeben würden?
 
 "@ -f $highLight2,$reset
 
@@ -202,20 +207,20 @@ $prompt {0}Get-Process{3} {1}pwsh,sys*{3} {2}-IncludeUsername{3}
 
 $P14 = @"
 
-The more time you spend reading help, the easier it will become.
+Je mehr Zeit Sie mit dem Lesen der Hilfe verbringen, desto einfacher wird es.
 
 "@
 
 $P15 = @"
-One last help item you need to know are the {0}about{4} topics. These are help files about
-PowerShell concepts and terminology. All of these topics begin with {1}about_{4}.
+Ein letztes Hilfethema, das Sie kennen sollten, sind die {0}about{4}-Themen. Dabei handelt es sich um
+Hilfedateien zu PowerShell‑Konzepten und -Terminologie. Alle diese Themen beginnen mit {1}about_{4}.
 
-Use the {2}help{4} command to display them. Wildcards are allowed and PowerShell will tab complete
-the topic name.
+Verwenden Sie den Befehl {2}help{4}, um sie anzuzeigen. Platzhalter sind erlaubt und PowerShell
+vervollständigt den Themennamen per Tab.
 
 $prompt {2}help{4} {3}about_parameters{4}
 
-Now, if you haven't done so in a while, go ahead run {2}Update-Help{4}.
+Wenn Sie das schon länger nicht getan haben, führen Sie jetzt {2}Update-Help{4} aus.
 
 "@ -f $highLight,$highLight2,$cmdStyle,$defaultTokenStyle,$reset
 #endregion